Work: move work controller to single file and do the remaining required refactor
